% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRANOVA.R
\name{anova_linear_regression}
\alias{anova_linear_regression}
\title{ANOVA for Linear Regression Models}
\usage{
anova_linear_regression(model, X, Y)
}
\arguments{
\item{model}{A list containing the components of a fitted linear regression model.
This should include the model coefficients, the matrix of predictors (X), and the vector of responses (Y).}

\item{X}{A numeric matrix of predictor variables.}

\item{Y}{A numeric vector of the response variable used in the linear regression model.}
}
\value{
A data frame with the ANOVA table, including the source of variation, sum of squares, degrees of freedom, mean squares, F value, and the associated p-value.
}
\description{
This function performs an Analysis of Variance (ANOVA) for a fitted linear regression model.
It computes the sum of squares due to regression (SSR), sum of squares error (SSE),
total sum of squares (SST), degrees of freedom, mean squares, F-statistic, and the p-value for the F-statistic.
}
\examples{
X <- matrix(rnorm(100), ncol = 5)
Y <- X \%*\% runif(5) + rnorm(20)
fit <- multivariate_linear_regression(X, Y)
anova_table <- anova_linear_regression(fit, X, Y)
print(anova_table)
}
