% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{covarianceMatrix}
\alias{covarianceMatrix}
\title{Compute Covariance Matrix}
\usage{
covarianceMatrix(x, y = NULL)
}
\arguments{
\item{x}{A numeric matrix.}

\item{y}{An optional second numeric matrix to compute the covariance with. If not provided, the function computes the covariance of `x`.}
}
\value{
A numeric matrix representing the covariance matrix.
}
\description{
This function computes the covariance matrix. It can compute the covariance matrix of a single matrix or the covariance between two matrices.
}
\examples{
mat1 <- matrix(1:4, nrow = 2)
mat2 <- matrix(1:6, nrow = 2)
covarianceMatrix(mat1)
covarianceMatrix(mat1, mat2)
}
