% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Test Dataset}
\format{
A data frame with 20 rows and 8 columns:
\describe{
  \item{X1}{X value}
  \item{X2}{X value}
  \item{X3}{X value}
  \item{X4}{X value}
  \item{Y1}{Y value}
  \item{Y2}{Y value}
  \item{Y3}{Y value}
  \item{Y4}{Y value}
}
}
\usage{
data
}
\description{
This dataset is used to test the cca_simple function.
}
\keyword{datasets}
