% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenValues}
\alias{eigenValues}
\title{Compute Eigenvalues}
\usage{
eigenValues(mat)
}
\arguments{
\item{mat}{A numeric matrix (arma::mat) for which to compute the eigenvalues.}
}
\value{
A numeric vector (arma::vec) containing the eigenvalues in descending order.
}
\description{
This function computes the eigenvalues of a given matrix. It returns the eigenvalues in descending order.
}
\examples{
mat <- matrix(c(2, -1, -1, 2), nrow = 2)
eigenValues(mat)
}
