% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{linear_regression_function}
\alias{linear_regression_function}
\title{Linear Regression Function}
\usage{
linear_regression_function(df, y, x)
}
\arguments{
\item{df}{A dataframe containing the variables of interest.}

\item{y}{The dependent variable from the dataframe.}

\item{x}{The covariates from the dataframe.}
}
\value{
A list containing data related to linear regression models.
}
\description{
This function performs a linear regression and calculates data related to linear regression models.
}
\examples{
data(mtcars)
result=linear_regression_function(mtcars, "mpg", c("cyl","disp","hp"))
}
