% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearModel.R
\name{linearModel}
\alias{linearModel}
\title{Linear Model}
\usage{
linearModel(formula, data, print.summary)
}
\arguments{
\item{formula}{a formula object that specifies the form of the model to be fitted.}

\item{data}{a data frame containing the variables to be used in the linear regression.}

\item{print.summary}{Boolean indicator for printing a summary table}
}
\value{
a list of results including coefficients, residuals, and fitted values.
}
\description{
The function linearModel is used to fit a simple or multiple linear regression model
}
\examples{
data <- iris
form <- as.formula(Sepal.Length ~ Sepal.Width + Petal.Width + Petal.Length + Species)
lm1 <- linearModel(formula = form, data = iris, print.summary = TRUE)

}
