% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_ec.R
\name{lm_ec}
\alias{lm_ec}
\title{Create linear regression model}
\usage{
lm_ec(formula, data)
}
\arguments{
\item{formula}{The formula of the linear regression model}

\item{data}{The data the model will be fit to}
}
\value{
The coefficients of the model
}
\description{
Create linear regression model
}
\examples{
data(mtcars)
lm_ec(mpg ~ hp + wt, data = mtcars)
}
