% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_summ_ec.R
\name{lm_summ_ec}
\alias{lm_summ_ec}
\title{Create model summary}
\usage{
lm_summ_ec(model)
}
\arguments{
\item{model}{The model output from lm_ec function}
}
\value{
A sumary of the residuals, coefficients and model fit
}
\description{
Create model summary
}
\examples{
data(mtcars)
m1 <- lm_ec(mpg ~ hp + wt, data = mtcars)
lm_summ_ec(m1)
}
