% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRM.R
\name{LRM}
\alias{LRM}
\title{LRM}
\usage{
LRM(formula, data, alpha = 0.05)
}
\arguments{
\item{formula}{It's similar to "formula" in the R function lm(): a symbolic description of the model to be fitted.}

\item{data}{A data frame, containing both the predictors of interests and outcome variables.}

\item{alpha}{A number represents the significance level, used to calculate confidence interval of estimated parameters (alpha=0.05 by default).}
}
\value{
The function returns the Coefficients' matrix with the estimated value, Standard error, t-statistics and p-value of each coefficient. It also automatically print the F statistics, R squared and adjusted R squared. Besides, it also returns the covariance matrix and confidence interval table
}
\description{
The LRM function will fit a linear regression model based on given data set. It shows the summary of coefficients, confidence interval and covariance table of estimated parameters of the model.
The function may not explore interaction between predictors and it cannot fit a model without an intercept properly.
The function can fit data with NA values.
}
\examples{
LRM(mpg~wt+drat,data=mtcars,alpha=0.01)
LRM(mpg~wt,data=mtcars,alpha=0.05)

}
