% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrixInverse}
\alias{matrixInverse}
\title{Matrix Inverse}
\usage{
matrixInverse(mat)
}
\arguments{
\item{mat}{NumericMatrix, the matrix for which to find the inverse.}
}
\value{
NumericMatrix, the inverse of \code{mat}.
}
\description{
Calculates the inverse of a matrix.
}
\examples{
mat <- matrix(1:4, nrow = 2)
matrixInverse(mat)
}
