% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrixMultiply}
\alias{matrixMultiply}
\title{Matrix Multiplication}
\usage{
matrixMultiply(mat1, mat2)
}
\arguments{
\item{mat1}{A numeric matrix.}

\item{mat2}{A numeric matrix.}
}
\value{
A numeric matrix that is the result of multiplying `mat1` by `mat2`.
}
\description{
Multiplies two matrices. The number of columns in the first matrix must match the number of rows in the second matrix.
}
\examples{
mat1 <- matrix(1:4, nrow = 2)
mat2 <- matrix(1:6, nrow = 2)
matrixMultiply(mat1, mat2)
}
