% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixsqrt.R
\name{matrixsqrt}
\alias{matrixsqrt}
\title{Matrix Square Root}
\usage{
matrixsqrt(m)
}
\arguments{
\item{m}{A matrix which we want to find the square root of.}
}
\value{
The square root of the input matrix m.
}
\description{
This function calculates the square root of a matrix.
}
\examples{
x <- cov(matrix(rnorm(2000), 1000, 2))
matrixsqrt(x)

}
