% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrixSqrtcpp}
\alias{matrixSqrtcpp}
\title{Compute the Square Root of a Matrix}
\usage{
matrixSqrtcpp(m)
}
\arguments{
\item{m}{A square numeric matrix (arma::mat).}
}
\value{
A numeric matrix (arma::mat) that is the square root of the input matrix.
}
\description{
This function computes the square root of a given square matrix using eigenvalue decomposition.
}
\examples{
mat <- matrix(c(4, 0, 0, 4), nrow = 2)
matrixSqrtcpp(mat)
}
