% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRANOVA.R
\name{multivariate_linear_regression}
\alias{multivariate_linear_regression}
\title{Multivariate Linear Regression}
\usage{
multivariate_linear_regression(X, Y)
}
\arguments{
\item{X}{A numeric matrix of predictor variables.}

\item{Y}{A numeric vector of the response variable.}
}
\value{
A list containing the regression coefficients, the residuals, the residual standard error, and the multiple R-squared statistic.
}
\description{
This function performs multivariate linear regression to predict the response variable given a set of predictors using the least squares method.
}
