% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{print.linear_regression}
\alias{print.linear_regression}
\title{Print Linear Regression Results}
\usage{
\method{print}{linear_regression}(x, ...)
}
\arguments{
\item{x}{An object of class linear_regression.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print Linear Regression Results
}
