% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Ridge regression}
\usage{
ridge_regression(y, X, k = 0)
}
\arguments{
\item{y}{is a matrix of output labels, each row represents an observation}

\item{x}{is a matrix of input features, each row represents an observation}
}
\value{
the parameters \beta for ridge regression
}
\description{
This function is designed to obtain the coefficient b from ridge regression \deqn{Y = X\beta + \lambda*I}
}
\examples{
square(3)

}
