% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{sampledata}
\alias{sampledata}
\title{Sample Dataset}
\format{
A data frame with 100 rows and 4 columns:
\describe{
  \item{X1}{Random variable from N(0,1), which is independent from each other}
  \item{X2}{Random variable from N(0,1), which is independent from each other}
  \item{X3}{Random variable from N(0,1), which is independent from each other}
  \item{Y}{Random variable from N(0,1), which is independent from each other}
}
}
\usage{
sampledata
}
\description{
A sample dataset included in the R package. It includes X = (X1,X2,X3), Y where each of them to be iid from N(0,1),
}
\keyword{datasets}
