% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaleC}
\alias{scaleC}
\title{Scale a Matrix}
\usage{
scaleC(x)
}
\arguments{
\item{x}{A numeric matrix.}
}
\value{
A scaled numeric matrix.
}
\description{
This function centers and scales a numeric matrix.
}
\examples{
scaleC(matrix(1:4, nrow = 2))
}
