% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solveMatrixcpp}
\alias{solveMatrixcpp}
\title{Compute the Inverse of a Matrix}
\usage{
solveMatrixcpp(x)
}
\arguments{
\item{x}{A numeric matrix (NumericMatrix).}
}
\value{
A numeric matrix (NumericMatrix) that is the inverse of the input matrix.
}
\description{
This function computes the inverse of a given numeric matrix using the Armadillo library.
}
\examples{
mat <- matrix(c(4, 7, 2, 6), nrow = 2)
solveMatrixcpp(mat)
}
