% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{svdDecomposition}
\alias{svdDecomposition}
\title{Singular Value Decomposition}
\usage{
svdDecomposition(X)
}
\arguments{
\item{X}{A numeric matrix (arma::mat) for which the SVD is to be computed.}
}
\value{
A list containing the matrices U, S, and V from the SVD.
}
\description{
This function performs a singular value decomposition on a given numeric matrix using the Armadillo library.
}
\examples{
mat <- matrix(c(1, 2, 3, 4), nrow = 2)
svdDecomposition(mat)
}
