% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test Dataset}
\format{
A data frame with 20 rows and 8 columns:
\describe{
  \item{column 1 to 4}{X value}
  \item{column 5 to 8}{Y value}
}
}
\usage{
test_data
}
\description{
This dataset is used to test the cca_simple function.
}
\keyword{datasets}
