% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ybanova.R
\name{ybanova}
\alias{ybanova}
\title{ANOVA table generated by yblm}
\arguments{
\item{model_results}{A list containing the output from the `yblm` function. It should be an model object completley from yblm function output.}
}
\value{
A data frame representing the ANOVA table with the following columns:
`Source` (source of variance), `Df` (degrees of freedom), `Sum Sq` (sum of squares),
`Mean Sq` (mean square), `F value`, and `Pr(>F)` (p-value for the F-statistic).
}
\description{
This function computes an ANOVA (Analysis of Variance) table for a simple linear regression model.
It is intended to be used with the output from the `yblm` function.
}
\examples{
#Start with the sample data given in the package
yblm(sampledata[,1:3],sampledata[,4]) -> fit_model
#The following will generate an ANOVA table to you
ybanova(fit_model)

#-------------or---------------------------
X = mvrnorm(100,mu = c(0,0),Sigma = matrix(c(1,.5,.5,1),2,2))
Y = rnorm(100,1)
yblm(X,Y,inter = TRUE) -> fit_model2
#The following will generate an ANOVA table to you
ybanova(fit_model2)
}
