% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRM.R
\name{LRM}
\alias{LRM}
\title{LRM}
\usage{
LRM(formula, data, alpha = 0.05)
}
\arguments{
\item{formula}{It's similar to "formula" in the origin R function lm(): a symbolic description of the model to be fitted.}

\item{data}{an data frame, containing the both the predictors and outcome variables in the model. If not found in data, the variables are taken from environment, typically the environment from which lm_s is called.}

\item{alpha}{a numeric number specifying the significance level to calculate confidence interval of estimated betas (0.05 by default).}
}
\value{
The function returns the Coefficients' matrix with the estimated value, Standard error, t-statistics and p-value of each coefficient. It also automatically print the F statistics, R squared and adjusted R squared. Besides, it also returns the variance matrix and confidence interval table.
}
\description{
The LRM function will fit a linear model using given data set. It will print summary of coefficients and confidence interval table of the model.
The function can treat NA with different action.
The function may not explore interaction among predictors and it cannot fit a model without an intercept properly.
}
\examples{
LRM(mpg~wt+drat,data=mtcars,alpha=0.01)
LRM(mpg~wt,data=mtcars,alpha=0.05)

}
