% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrixMultiply}
\alias{matrixMultiply}
\title{Matrix Multiplication}
\usage{
matrixMultiply(mat1, mat2)
}
\arguments{
\item{mat1}{NumericMatrix, the first matrix.}

\item{mat2}{NumericMatrix, the second matrix.}
}
\value{
NumericMatrix, the result of multiplying \code{mat1} and \code{mat2}.
}
\description{
Multiplies two matrices.
}
\examples{
mat1 <- matrix(1:4, nrow = 2)
mat2 <- matrix(1:4, nrow = 2)
matrixMultiply(mat1, mat2)
}
