% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{print.linear_regression}
\alias{print.linear_regression}
\title{Print Linear Regression Results}
\usage{
\method{print}{linear_regression}(object)
}
\arguments{
\item{object}{An object of class linear_regression.}
}
\description{
Print Linear Regression Results
}
