% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.linear_regression}
\alias{print.linear_regression}
\title{Print Linear Regression Results}
\usage{
\method{print}{linear_regression}(x, ...)
}
\arguments{
\item{x}{An object of class linear_regression.}

\item{...}{Additional arguments (not used).}
}
\description{
This function prints a summary of the results from the linear_regression function.
}
