% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Ridge Regression}
\usage{
ridge_regression(y, X, k = 0)
}
\arguments{
\item{y}{is a matrix of output labels, each row represents an observation}

\item{k}{is a predetermined adjusting number. When k=0, model becomes linear regression}

\item{x}{is a matrix of input features, each row represents an observation}
}
\value{
the matrix of parameters \eqn{\beta} for ridge regression
}
\description{
This function is designed to obtain the coefficient b from ridge regression
\deqn{Y = X\beta + e}
and minimize \eqn{||Y-X\beta||^2+\lambda||\beta||},
where \eqn{\lambda} served as a predetermined hyperparamter to shift the diagonals of moment matrix.
More information about this method could be found in references.
}
\examples{
square(3)

}
\references{
Wikipedia contributors. (n.d.). Ridge regression. In Wikipedia. Retrieved [November, 11th, 2023], from https://en.wikipedia.org/wiki/Ridge_regression
}
